Fix firmas factory

+++ database/migrations/tenant/2024_02_21_103918_create_pazysalvos_firma.php
@@ -12,6 +12,30 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (! Schema::hasTable('firmaspdf')) {
+            Schema::create('firmaspdf', function (Blueprint $table) {
+                $table->id('idfirmas');
+                $table->string('descripcion', 100);
+                $table->string('modulo', 100);
+                $table->char('estado', 1);
+            });
+        }
+
+        if (! Schema::hasTable('firmaspdf_det')) {
+            Schema::create('firmaspdf_det', function (Blueprint $table) {
+                $table->id('iddet');
+                $table->integer('idfirmas');
+                $table->string('documento', 100);
+                $table->string('funcionario', 100);
+                $table->integer('idcargo');
+                $table->string('nomcargo', 100);
+                $table->integer('orden');
+                $table->char('estado', 1);
+                $table->string('fecha', 10);
+                $table->string('titulo', 100);
+            });
+        }
+
         DB::table('firmaspdf')->insert([
             'idfirmas' => 8,
             'descripcion' => 'PAZ Y SALVO PREDIAL',
@@ -37,7 +61,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        DB::table('firmaspdf_det')->where('idfirmas', 8)->delete();
-        DB::table('firmaspdf')->where('idfirmas', 8)->delete();
+        if (Schema::hasTable('firmaspdf_det')) {
+            DB::table('firmaspdf_det')->where('idfirmas', 8)->delete();
+        }
+
+        if (Schema::hasTable('firmaspdf')) {
+            DB::table('firmaspdf')->where('idfirmas', 8)->delete();
+        }
     }
 };
